### 2025-08-04 7cde70b
Fix DOI fields
