+++ paper/paper.bib
@@ -44,7 +44,7 @@
   author       = {Hill, Alan P and Young, Robert J},
   url          = {https://www.sciencedirect.com/science/article/pii/S1359644610001923},
   date         = {2010},
-  doi          = {https://doi.org/10.1016/j.drudis.2010.05.016},
+  doi          = {10.1016/j.drudis.2010.05.016},
   issn         = {1359-6446},
   journaltitle = {Drug Discov. Today},
   number       = {15},
@@ -241,7 +241,7 @@
   author       = {Zhang, Haohui and Wu, Juntong and Liu, Shichao and Han, Shen},
   url          = {https://www.sciencedirect.com/science/article/pii/S1566253523004086},
   date         = {2024},
-  doi          = {https://doi.org/10.1016/j.inffus.2023.102092},
+  doi          = {10.1016/j.inffus.2023.102092},
   issn         = {1566-2535},
   journaltitle = {Information Fusion},
   keywords     = {Molecular property prediction,Graph neural networks,Multi-modal fusion,Unsupervised pre-training},
@@ -295,7 +295,7 @@
   author       = {Gadiya, Yojana and Gribbon, Philip and Hofmann-Apitius, Martin and Zaliani, Andrea},
   url          = {https://www.sciencedirect.com/science/article/pii/S2667318523000132},
   date         = {2023},
-  doi          = {https://doi.org/10.1016/j.ailsci.2023.100069},
+  doi          = {10.1016/j.ailsci.2023.100069},
   issn         = {2667-3185},
   journaltitle = {Artificial Intelligence in the Life Sciences},
   keywords     = {Patent analysis,Drug discovery,Patent landscape,Rare disease,Alzheimer's disease,Drug repurposing,Target importance},
