### 2025-08-04 dcbacd9
Add JOSS badge
