### 2025-08-05 #8712 @jhylin
@espottesmith Hi, thanks for asking but I don't feel I'm experienced enough yet to review this paper/codebase. I think it'll be better for Pat, Greg or others to do this if possible. (@cthoyt thanks for mentioning me in the acknowledgements though!)
