+++ docs/source/conf.py
@@ -243,6 +243,8 @@ texinfo_documents = [
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3", None),
     "pandas": ("https://pandas.pydata.org/docs", None),
+    "chemfp": ("https://chemfp.readthedocs.io/en/latest", None),
+    "pystow": ("https://pystow.readthedocs.io/en/latest", None),
 }
 
 autoclass_content = "both"
