+++ docs/source/reference.rst
@@ -1,0 +1,5 @@
+Reference
+=========
+
+.. automodule:: chembl_downloader
+    :members:
