+++ docs/source/usage.rst
@@ -1,5 +1,0 @@
-Usage
-=====
-
-.. automodule:: chembl_downloader.api
-    :members:
