+++ docs/source/queries.rst
@@ -1,0 +1,6 @@
+SQL Queries
+===========
+
+.. automodapi:: chembl_downloader.queries
+    :no-heading:
+    :include-all-objects:
