+++ docs/source/fingerprints.rst
@@ -25,5 +25,6 @@ The following wraps the :func:`chembl_downloader.download_fps` function with
     arena = chembl_downloader.chemfp_load_fps()
 
 The ``version`` and other keyword arguments are also valid for this function. More
-information on working with the :class:`chemfp.arena.FingerprintArena` object can be found `here
+information on working with the :class:`chemfp.arena.FingerprintArena` object can be
+found `here
 <https://chemfp.readthedocs.io/en/latest/using-api.html#working-with-a-fingerprintarena>`_.
