+++ docs/source/extras.rst
@@ -1,0 +1,5 @@
+Extras
+======
+
+.. automodapi:: chembl_downloader.contrib
+    :no-heading:
