### 2025-08-12 #8712 @espottesmith
@HzaCode @OlivierBeq @iwatobipen

Can any of y'all review this paper/codebase?
