### 2025-08-12 #8712 @HzaCode
Hi @espottesmith ,

Thanks for the invitation.

This is a bit of a tricky one, isn't it? As you know, we all just interacted on my manuscript. Because of that, I'm honestly a little concerned I wouldn't be able to give the author the tough, critical review that's actually helpful for improving their paper.

I really think someone with fresh eyes would be a better fit and do a better service to the author. So I'll have to pass this time.

Hope that makes sense!
