### 2025-08-21 #8844 @editorialbot
[REVIEW]: Improving reproducibility of cheminformatics workflows with `chembl-downloader`
**Submitting author:** @cthoyt (<a href="http://orcid.org/0000-0003-4423-4370">Charles Tapley Hoyt</a>)
**Repository:** https://github.com/cthoyt/chembl-downloader
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.2
