### 2025-08-22 1bed24a
Improve version handling
This makes it much easier to work in a loop
