+++ src/chembl_downloader/api.py
@@ -15,6 +15,7 @@ from contextlib import closing, contextmanager
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal, NamedTuple, overload
 from xml.etree import ElementTree
+from typing_extensions import TypeAlias
 
 import pystow
 from tqdm import tqdm
@@ -56,6 +57,7 @@ __all__ = [
     "query",
     "supplier",
     "versions",
+    "VersionHint",
 ]
 
 logger = logging.getLogger(__name__)
@@ -64,6 +66,9 @@ logger = logging.getLogger(__name__)
 PYSTOW_PARTS = ["chembl"]
 RELEASE_PREFIX = "* Release:"
 DATE_PREFIX = "* Date:"
+
+#: A hint for a version, which can either be an integer or string
+VersionHint: TypeAlias = str | int
 
 
 class VersionPathPair(NamedTuple):
@@ -75,7 +80,7 @@ class VersionPathPair(NamedTuple):
 
 def _removeprefix(s: str, prefix: str) -> str:
     if s.startswith(prefix):
-        return s[len(prefix) :]
+        return s[len(prefix):]
     return s
 
 
@@ -107,7 +112,7 @@ def versions() -> list[str]:
 
 def _download_helper(
     suffix: str,
-    version: str | None = None,
+    version: VersionHint | None = None,
     prefix: Sequence[str] | None = None,
     *,
     return_version: bool,
@@ -132,9 +137,7 @@ def _download_helper(
     if version is None:
         version = latest()
 
-    # for versions 22.1 and 24.1, it's important to canonicalize the version number
-    # for versions < 10 it's important to left pad with a zero
-    fmt_version = version.replace(".", "_").zfill(2)
+    fmt_version, version = _clean_version(version)
 
     base = f"ftp://ftp.ebi.ac.uk/pub/databases/chembl/ChEMBLdb/releases/chembl_{fmt_version}"
     if filename_repeats_version:
@@ -156,10 +159,22 @@ def _download_helper(
     raise ValueError(f"could not find {filename} in data for ChEMBL {fmt_version} in {base}")
 
 
+def _clean_version(version: VersionHint) -> tuple[str, str]:
+    if isinstance(version, int):
+        # versions 1-9 are left padded with a zero
+        version = f"{version:02}"
+
+    # for versions 22.1 and 24.1, it's important to canonicalize the version number
+    # for versions < 10 it's important to left pad with a zero
+    fmt_version = version.replace(".", "_").zfill(2)
+
+    return fmt_version, version
+
+
 # docstr-coverage:excused `overload`
 @overload
 def download_sqlite(
-    version: str | None = ...,
+    version: VersionHint | None = ...,
     *,
     prefix: Sequence[str] | None = ...,
     return_version: Literal[True] = ...,
@@ -169,7 +184,7 @@ def download_sqlite(
 # docstr-coverage:excused `overload`
 @overload
 def download_sqlite(
-    version: str | None = ...,
+    version: VersionHint | None = ...,
     *,
     prefix: Sequence[str] | None = ...,
     return_version: Literal[False] = ...,
@@ -177,7 +192,7 @@ def download_sqlite(
 
 
 def download_sqlite(
-    version: str | None = None,
+    version: VersionHint | None = None,
     *,
     prefix: Sequence[str] | None = None,
     return_version: bool = False,
@@ -204,7 +219,7 @@ def download_sqlite(
 # docstr-coverage:excused `overload`
 @overload
 def download_extract_sqlite(
-    version: str | None = ...,
+    version: VersionHint | None = ...,
     *,
     prefix: Sequence[str] | None = ...,
     return_version: Literal[True] = ...,
[…]
