### 2025-08-22 328afc9
Add query unpacking support
