+++ src/chembl_downloader/cli.py
@@ -11,6 +11,7 @@ from .api import (
     get_date,
     get_substructure_library,
     query,
+    query_one,
     versions,
 )
 from .queries import ACTIVITIES_QUERY, ID_NAME_QUERY
@@ -40,6 +41,10 @@ def download(version: str | None) -> None:
 @verbose_option  # type:ignore
 def test(version: str | None) -> None:
     """Run test queries."""
+    click.secho("\n\nNumber of Activities\n", fg="green")
+    count = query_one("SELECT COUNT(activity_id) FROM activities", version=version)
+    click.echo(f"{count:,}\n")
+
     click.secho("ID to Name Query\n", fg="green")
     df = query(ID_NAME_QUERY + "\nLIMIT 5", version=version)
     click.echo(df.to_markdown(index=False))
@@ -83,7 +88,7 @@ def _count_compounds(version: str) -> str:
     from .queries import COUNT_QUERY_SQL
 
     try:
-        total_compounds = query(COUNT_QUERY_SQL, version=version)["count"][0]
+        total_compounds = query_one(COUNT_QUERY_SQL, version=version)
     except Exception:
         return "-"
     return f"{total_compounds:,}"
