+++ tests/test_api.py
@@ -1,13 +1,56 @@
 """Tests for the API."""
 
+import shutil
+import sqlite3
+import tempfile
 import unittest
+from pathlib import Path
 from textwrap import dedent
 
+import pandas as pd
+from pystow.constants import PYSTOW_HOME_ENVVAR
+from pystow.utils import mock_envvar, open_tarfile
+
 import chembl_downloader
+
+TV = "26"
 
 
 class TestApi(unittest.TestCase):
     """Tests for the API."""
+
+    def setUp(self) -> None:
+        """Set up the test case."""
+        rows = [(1,), (2,), (3,)]
+        columns = ["activity_id"]
+        self.directory_obj = tempfile.TemporaryDirectory()
+        self.directory = Path(self.directory_obj.name)
+        data_path = self.directory.joinpath("test.db")
+        self.tarfile_path = self.directory.joinpath(f"chembl_{TV}_sqlite.tar.gz")
+
+        df = pd.DataFrame(rows, columns=columns)
+        with sqlite3.connect(data_path) as conn:
+            df.to_sql("activities", conn)
+
+        inner_path = f"data/chembl_{TV}/chembl_{TV}_sqlite/chembl_{TV}.db"
+        with (
+            open_tarfile(self.tarfile_path, inner_path, operation="write") as file,
+            data_path.open("rb") as f2,
+        ):
+            shutil.copyfileobj(f2, file)
+
+    def tearDown(self) -> None:
+        """Tear down the test case."""
+        self.directory_obj.cleanup()
+
+    def test_query(self) -> None:
+        """Test querying."""
+        with mock_envvar(PYSTOW_HOME_ENVVAR, self.directory.as_posix()):
+            chembl_downloader.query_one(
+                "SELECT COUNT(activity_id) FROM activities",
+                prefix=[],
+                version=TV,
+            )
 
     def test_latest_version(self) -> None:
         """Test getting the latest version."""
