+++ src/chembl_downloader/__init__.py
@@ -23,7 +23,7 @@ from .api import (
     iterate_smiles,
     latest,
     query,
-    query_one,
+    query_scalar,
     supplier,
     versions,
 )
@@ -51,7 +51,7 @@ __all__ = [
     "latest",
     "queries",
     "query",
-    "query_one",
+    "query_scalar",
     "supplier",
     "versions",
 ]
