+++ src/chembl_downloader/queries.py
@@ -13,6 +13,10 @@ __all__ = [
     "CHEBI_UNMAPPED_SQL",
     "DRUG_INDICATIONS_SQL",
     "ID_NAME_QUERY",
+    'COUNT_ASSAYS_SQL',
+    "COUNT_ACTIVITIES_SQL",
+    "COUNT_COMPOUNDS_SQL",
+    "COUNT_QUERY_SQL",
     # Functions
     "get_assay_sql",
     "get_document_molecule_sql",
@@ -191,3 +195,12 @@ def get_document_molecule_sql(document_chembl_id: str) -> str:
             WHERE DOCS.chembl_id = '{document_chembl_id}'
         """  # noqa: S608
     )
+
+#: Count the number of activity records
+COUNT_ACTIVITIES_SQL = "SELECT COUNT(*) from activities"
+
+#: Count the number of assay records
+COUNT_ASSAYS_SQL = "SELECT COUNT(*) from assays"
+
+#: Count the number of compound records
+COUNT_COMPOUNDS_SQL = "SELECT COUNT(*) from molecule_dictioanry"
