### 2025-08-23 41a95fe
Add shortcut for querying for a scalar (#23)
