+++ src/chembl_downloader/api.py
@@ -25,6 +25,7 @@ if TYPE_CHECKING:
     import pandas
     import rdkit.Chem
     import rdkit.Chem.rdSubstructLibrary
+    from numpy.typing import NDArray
 
 __all__ = [
     "VersionPathPair",
@@ -54,6 +55,7 @@ __all__ = [
     "iterate_smiles",
     "latest",
     "query",
+    "query_scalar",
     "supplier",
     "versions",
 ]
@@ -141,19 +143,28 @@ def _download_helper(
         filename = f"chembl_{fmt_version}{suffix}"
     else:
         filename = suffix
+
+    if prefix is None:
+        prefix = PYSTOW_PARTS
+
+    module = pystow.module(*prefix, fmt_version)
+
     for url in [
         f"{base}/{filename}",
         f"{base}/archived/{filename}",
     ]:
         try:
-            path = pystow.ensure(*(prefix or PYSTOW_PARTS), fmt_version, url=url)
+            path = module.ensure(url=url)
         except OSError:
             continue
         if return_version:
             return VersionPathPair(version, path)
         else:
             return path
-    raise ValueError(f"could not find {filename} in data for ChEMBL {fmt_version} in {base}")
+    raise ValueError(
+        f"could not find {filename} in data for ChEMBL {fmt_version} in {base} "
+        f"with PyStow module at {module.base}"
+    )
 
 
 # docstr-coverage:excused `overload`
@@ -243,7 +254,9 @@ def download_extract_sqlite(
         directories
     """
     if version is not None:
-        _directory = pystow.join(*(prefix or PYSTOW_PARTS), version)
+        if prefix is None:
+            prefix = PYSTOW_PARTS
+        _directory = pystow.join(*prefix, version)
         if _directory.is_dir():
             rv = _find_sqlite_file(_directory)
             if rv:
@@ -355,8 +368,8 @@ def query(
 
     Example: .. code-block:: python
 
-        import chembl_downloader from chembl_downloader.queries import
-        ID_NAME_QUERY_EXAMPLE
+        import chembl_downloader
+        from chembl_downloader.queries import ID_NAME_QUERY_EXAMPLE
 
         df = chembl_downloader.query(ID_NAME_QUERY_EXAMPLE)
     """
@@ -364,6 +377,38 @@ def query(
 
     with connect(version=version, prefix=prefix) as con:
         return pd.read_sql(sql, con=con, **kwargs)
+
+
+def query_scalar(
+    sql: str,
+    version: str | None = None,
+    *,
+    prefix: Sequence[str] | None = None,
+    **kwargs: Any,
+) -> Any:
+    """Ensure the data is available, run the query, then extract the result.
+
+    Similar to :func:`query`, but automatically unpacks the value, assuming that only
+    one record is returned with just a single column.
+
+    :param sql: A SQL query string or table name
+    :param version: The version number of ChEMBL to get. If none specified, uses
+        :func:`latest` to look up the latest.
+    :param prefix: The directory inside :mod:`pystow` to use
+    :param kwargs: keyword arguments to pass through to :func:`pandas.read_sql`, such as
+        ``index_col``.
+
+    :returns: A value (int, str, etc.) from the database.
+
+    Example: .. code-block:: python
+
+        import chembl_downloader
+
+        sql = "SELECT COUNT(activity_id) FROM activities"
+        count: int = chembl_downloader.query_one(sql)
+    """
+    df = query(sql, version=version, prefix=prefix, **kwargs)
+    return df[df.columns[0]][0]
 
 
 # docstr-coverage:excused `overload`
@@ -434,7 +479,7 @@ def iterate_fps(
     *,
[…]
