+++ tox.ini
@@ -45,6 +45,7 @@ commands =
 extras =
     # See the [project.optional-dependencies] entry in pyproject.toml for "tests"
     tests
+    pandas
 set_env =
     # this setting gets inherited into all environments, meaning
     # that things that call uv commands don't require a --preview
