+++ tox.ini
@@ -128,6 +128,7 @@ description = Run the mypy tool to check static typing on the project. Installs 
 deps =
     mypy
     types-tabulate
+    types-requests
 commands = mypy --ignore-missing-imports --strict src/ tests/
 
 [testenv:doc8]
