### 2025-08-23 f6aa725
Improve handling of versions (#22)
