+++ src/chembl_downloader/api.py
@@ -11,7 +11,7 @@ import tarfile
 from collections.abc import Generator, Iterable, Sequence
 from contextlib import closing, contextmanager
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Literal, NamedTuple, overload
+from typing import TYPE_CHECKING, Any, Literal, NamedTuple, TypeAlias, overload
 from xml.etree import ElementTree
 
 import pystow
@@ -27,6 +27,7 @@ if TYPE_CHECKING:
     from numpy.typing import NDArray
 
 __all__ = [
+    "VersionHint",
     "VersionPathPair",
     "chemfp_load_fps",
     "connect",
@@ -65,6 +66,9 @@ logger = logging.getLogger(__name__)
 PYSTOW_PARTS = ["chembl"]
 RELEASE_PREFIX = "* Release:"
 DATE_PREFIX = "* Date:"
+
+#: A hint for a version, which can either be an integer, string, or float (for minor versions)
+VersionHint: TypeAlias = str | int | float
 
 
 class VersionPathPair(NamedTuple):
@@ -107,9 +111,12 @@ def versions() -> list[str]:
     return sorted(version_list, reverse=True)
 
 
+_CHEMBL_HOST = "ftp.ebi.ac.uk/pub/databases/chembl/ChEMBLdb/releases"
+
+
 def _download_helper(
     suffix: str,
-    version: str | None = None,
+    version: VersionHint | None = None,
     prefix: Sequence[str] | None = None,
     *,
     return_version: bool,
@@ -131,46 +138,93 @@ def _download_helper(
 
     :raises ValueError: If file could not be downloaded
     """
-    if version is None:
-        version = latest()
+    version_info = _get_version_info(version, prefix=prefix)
 
-    # for versions 22.1 and 24.1, it's important to canonicalize the version number
-    # for versions < 10 it's important to left pad with a zero
-    fmt_version = version.replace(".", "_").zfill(2)
-
-    base = f"ftp://ftp.ebi.ac.uk/pub/databases/chembl/ChEMBLdb/releases/chembl_{fmt_version}"
     if filename_repeats_version:
-        filename = f"chembl_{fmt_version}{suffix}"
+        filename = f"chembl_{version_info.fmt_version}{suffix}"
     else:
         filename = suffix
 
-    if prefix is None:
-        prefix = PYSTOW_PARTS
-
-    module = pystow.module(*prefix, fmt_version)
-
-    for url in [
+    base = f"ftp://{_CHEMBL_HOST}/chembl_{version_info.fmt_version}"
+    urls = [
         f"{base}/{filename}",
         f"{base}/archived/{filename}",
-    ]:
+    ]
+    for url in urls:
         try:
-            path = module.ensure(url=url)
+            path = version_info.module.ensure(url=url)
         except OSError:
             continue
         if return_version:
-            return VersionPathPair(version, path)
+            return VersionPathPair(version_info.version, path)
         else:
             return path
-    raise ValueError(
-        f"could not find {filename} in data for ChEMBL {fmt_version} in {base} "
-        f"with PyStow module at {module.base}"
-    )
+
+    urls_fmt = "\n".join(f"   - {url}" for url in urls)
+    raise ValueError(f"""\
+
+[ChEMBL v{version_info.fmt_version}] could not ensure {filename}
+
+1. It wasn't already cached in the PyStow directory:
+   {version_info.module.base}
+
+2. It couldn't be downloaded from any of the following URLs:
+{urls_fmt}
+    """)
+
+
+class VersionInfo(NamedTuple):
+    """A pair of format version and regular version."""
+
+    fmt_version: str
+    version: str
+    module: pystow.Module
+
+
[…]
