### 2025-08-23 0fadadc
Replace internal removeprefix function
