### 2025-08-23 4026147
Replace internal removeprefix function (#25)
