### 2025-08-23 4a4bcc4
Improve summarize workflow
