### 2025-08-23 9efc6cc
Switch downloads to HTTPS (#27)
Goodbye FTP, thanks for all the fish
