+++ src/chembl_downloader/api.py
@@ -288,6 +288,7 @@ def download_extract_sqlite(
     *,
     prefix: Sequence[str] | None = None,
     return_version: bool = False,
+    retain: bool = False,
 ) -> Path | VersionPathPair:
     """Ensure the latest ChEMBL SQLite dump is downloaded and extracted.
 
@@ -296,6 +297,7 @@ def download_extract_sqlite(
     :param prefix: The directory inside :mod:`pystow` to use
     :param return_version: Should the version get returned? Turn this to true if you're
         looking up the latest version and want to reduce redundant code.
+    :param retain: If true, keeps the original archive.
 
     :returns: If ``return_version`` is true, return a pair of the version and the local
         file path to the downloaded ChEMBLSQLite database file. Otherwise, just return
@@ -306,18 +308,24 @@ def download_extract_sqlite(
     """
     version_info = _get_version_info(version, prefix)
 
-    name = f"chembl_{version_info.fmt_version}_sqlite.tar.gz"
+    name = f"chembl_{version_info.fmt_version}.db"
     rv = version_info.module.join(name=name)
 
     if not rv.is_file():
         # TODO make all versions accept VersionInfo
-        tar_path = download_sqlite(version=version_info.version, prefix=prefix, return_version=False)
+        tar_path = download_sqlite(
+            version=version_info.version, prefix=prefix, return_version=False
+        )
         with tarfile.open(tar_path, mode="r", encoding="utf-8") as tar_file:
             tar_info = _get_tar_info(tar_file)
             if tar_info is None:
                 raise FileNotFoundError("could not find a .db file in the ChEMBL archive")
-            logger.info("unarchiving %s", tar_path)
-            tar_file.extract(tar_info, rv)
+            logger.info("unarchiving %s to %s", tar_path, rv)
+            tar_file._extract_member(tar_info, rv.as_posix())
+
+        if not retain:
+            logger.info("deleting original archive %s", tar_path)
+            tar_path.unlink()
 
     if return_version:
         return VersionPathPair(version, rv)
