### 2025-08-23 f11500a
Refactor logic for extracting SQLite database from a TAR archive (#28)
