+++ src/chembl_downloader/api.py
@@ -270,6 +270,7 @@ def download_extract_sqlite(
     *,
     prefix: Sequence[str] | None = ...,
     return_version: Literal[True] = ...,
+    retain: bool = ...,
 ) -> VersionPathPair: ...
 
 
@@ -280,6 +281,7 @@ def download_extract_sqlite(
     *,
     prefix: Sequence[str] | None = ...,
     return_version: Literal[False] = ...,
+    retain: bool = ...,
 ) -> Path: ...
 
 
@@ -288,6 +290,7 @@ def download_extract_sqlite(
     *,
     prefix: Sequence[str] | None = None,
     return_version: bool = False,
+    retain: bool = False,
 ) -> Path | VersionPathPair:
     """Ensure the latest ChEMBL SQLite dump is downloaded and extracted.
 
@@ -296,6 +299,7 @@ def download_extract_sqlite(
     :param prefix: The directory inside :mod:`pystow` to use
     :param return_version: Should the version get returned? Turn this to true if you're
         looking up the latest version and want to reduce redundant code.
+    :param retain: If true, keeps the original archive.
 
     :returns: If ``return_version`` is true, return a pair of the version and the local
         file path to the downloaded ChEMBLSQLite database file. Otherwise, just return
@@ -304,36 +308,39 @@ def download_extract_sqlite(
     :raises FileNotFoundError: If no database file could be found in the extracted
         directories
     """
-    if version is not None:
-        version_info = _get_version_info(version, prefix)
-        _directory = version_info.module.base
-        if _directory.is_dir():
-            rv = _find_sqlite_file(_directory)
-            if rv:
-                if return_version:
-                    return VersionPathPair(version_info.version, rv)
-                return rv
+    version_info = _get_version_info(version, prefix)
 
-    version, path = download_sqlite(version=version, prefix=prefix, return_version=True)
+    name = f"chembl_{version_info.fmt_version}.db"
+    rv = version_info.module.join(name=name)
 
-    # Extraction will be done in the same directory as the download.
-    # All ChEMBL SQLite dumps have the same internal folder structure,
-    # so assume there's going to be a directory here
-    directory = path.parent.joinpath("data")
-    if not directory.is_dir():
-        logger.info("unarchiving %s to %s", path, directory)
-        with tarfile.open(path, mode="r", encoding="utf-8") as tar_file:
-            tar_file.extractall(directory)  # noqa:S202
-    else:
-        logger.debug("did not re-unarchive %s to %s", path, directory)
+    if not rv.is_file():
+        # TODO make all versions accept VersionInfo
+        tar_path = download_sqlite(
+            version=version_info.version, prefix=prefix, return_version=False
+        )
+        with tarfile.open(tar_path, mode="r", encoding="utf-8") as tar_file:
+            tar_info = _get_tar_info(tar_file)
+            if tar_info is None:
+                raise FileNotFoundError("could not find a .db file in the ChEMBL archive")
+            logger.info("unarchiving %s to %s", tar_path, rv)
+            tar_file._extract_member(tar_info, rv.as_posix())
 
-    rv = _find_sqlite_file(directory)
-    if rv is None:
-        raise FileNotFoundError("could not find a .db file in the ChEMBL archive")
-    elif return_version:
-        return VersionPathPair(version, rv)
+        if not retain:
+            logger.info("deleting original archive %s", tar_path)
+            tar_path.unlink()
+
+    if return_version:
+        return VersionPathPair(version_info.version, rv)
     else:
         return rv
+
+
+def _get_tar_info(tar_file: tarfile.TarFile) -> tarfile.TarInfo | None:
+    """Walk an archive and find a file with the ``.db`` extension."""
[…]
