### 2025-08-23 a2ca828
Enable passing version info
