### 2025-08-23 6ff1a36
Enable passing VersionInfo objects to version arguments (#29)
