+++ README.md
@@ -303,41 +303,45 @@ See
 files are not available for all versions. For example, the SQLite version of the
 database was first added in release 21 (2015-02-12).
 
-| ChEMBL Version | Release Date | Total Named Compounds _from SQLite_ |
-| -------------- | ------------ | ----------------------------------: |
-| 31             | 2022-07-12   |                              41,585 |
-| 30             | 2022-02-22   |                              41,549 |
-| 29             | 2021-07-01   |                              41,383 |
-| 28             | 2021-01-15   |                              41,049 |
-| 27             | 2020-05-18   |                              40,834 |
-| 26             | 2020-02-14   |                              40,822 |
-| 25             | 2019-02-01   |                              39,885 |
-| 24_1           | 2018-05-01   |                              39,877 |
-| 24             |              |                                     |
-| 23             | 2017-05-18   |                              39,584 |
-| 22_1           | 2016-11-17   |                                     |
-| 22             |              |                              39,422 |
-| 21             | 2015-02-12   |                              39,347 |
-| 20             | 2015-02-03   |                                   - |
-| 19             | 2014-07-23   |                                   - |
-| 18             | 2014-04-02   |                                   - |
-| 17             | 2013-09-16   |                                   - |
-| 16             | 2013-05-15   |                                   - |
-| 15             | 2013-01-30   |                                   - |
-| 14             | 2012-07-18   |                                   - |
-| 13             | 2012-02-29   |                                   - |
-| 12             | 2011-11-30   |                                   - |
-| 11             | 2011-06-07   |                                   - |
-| 10             | 2011-06-07   |                                   - |
-| 09             | 2011-01-04   |                                   - |
-| 08             | 2010-11-05   |                                   - |
-| 07             | 2010-09-03   |                                   - |
-| 06             | 2010-09-03   |                                   - |
-| 05             | 2010-06-07   |                                   - |
-| 04             | 2010-05-26   |                                   - |
-| 03             | 2010-04-30   |                                   - |
-| 02             | 2009-12-07   |                                   - |
-| 01             | 2009-10-28   |                                   - |
+| Version | Date       | Compounds |    Assays | Activities | Named Compounds |
+|---------|------------|----------:|----------:|-----------:|----------------:|
+| 35      | 2024-12-01 | 2,496,335 | 1,740,546 | 21,123,501 |          42,231 |
+| 34      | 2024-03-28 | 2,431,025 | 1,644,390 | 20,772,701 |          42,387 |
+| 33      | 2023-05-31 | 2,399,743 | 1,610,596 | 20,334,684 |          41,923 |
+| 32      | 2023-01-26 | 2,354,965 | 1,536,903 | 20,038,828 |          41,923 |
+| 31      | 2022-07-12 | 2,331,700 | 1,498,681 | 19,780,369 |          41,585 |
+| 30      | 2022-02-22 | 2,157,379 | 1,458,215 | 19,286,751 |          41,549 |
[…]
