### 2025-08-24 24f90ab
Add outputs
