### 2025-08-24 03f390a
Automate generating temporal summaries (#26)
