### 2025-08-26 #8844 @espottesmith
Hey @PatWalters, @dhimmel have y'all had a chance to begin looking at this code/paper?
