### 2025-08-26 7daceb8
Extend historical analysis
- fix old table names
- more tables/summary statistics
