### 2025-08-26 5873c16
Automate resizing
