### 2025-08-26 8d812bd
Extend historical analysis (#30)
