+++ src/chembl_downloader/cli.py
@@ -1,30 +1,27 @@
 """CLI for :mod:`chembl_downloader`."""
 
+from functools import partial
 from pathlib import Path
 
 import click
 from more_click import verbose_option
 from tqdm import tqdm
+from tqdm.contrib.concurrent import thread_map
 
 from .api import (
     SummaryTuple,
+    VersionHint,
+    VersionInfo,
     _get_version_info,
     download_extract_sqlite,
     download_readme,
     get_substructure_library,
     latest,
     query,
-    query_scalar,
     summarize,
     versions,
 )
-from .queries import (
-    ACTIVITIES_QUERY,
-    COUNT_ACTIVITIES_SQL,
-    COUNT_ASSAYS_SQL,
-    COUNT_COMPOUNDS_SQL,
-    ID_NAME_QUERY,
-)
+from .queries import ACTIVITIES_QUERY, ID_NAME_QUERY
 
 __all__ = [
     "main",
@@ -46,30 +43,39 @@ def download(version: str | None) -> None:
     click.echo(download_extract_sqlite(version=version))
 
 
-@main.command()
+@main.command(name="summarize")
 @version_option
 @verbose_option  # type:ignore
-def test(version: str | None) -> None:
-    """Run test queries."""
+def print_summary(version: str | None) -> None:
+    """Run test queries and print results to the console."""
+    summary = summarize(version=version)
+
     click.secho("Number of Activities", fg="green")
-    count = query_scalar(COUNT_ACTIVITIES_SQL, version=version)
-    click.echo(f"{count:,}")
+    click.echo(f"{summary.activities:,}")
 
     click.secho("\nNumber of Compounds", fg="green")
-    count = query_scalar(COUNT_COMPOUNDS_SQL, version=version)
-    click.echo(f"{count:,}")
+    click.echo(f"{summary.compounds:,}")
 
     click.secho("\nNumber of Assays", fg="green")
-    count = query_scalar(COUNT_ASSAYS_SQL, version=version)
-    click.echo(f"{count:,}\n")
+    click.echo(f"{summary.assays:,}\n")
 
-    click.secho("ID to Name Query\n", fg="green")
-    df = query(ID_NAME_QUERY + "\nLIMIT 5", version=version)
-    click.echo(df.to_markdown(index=False))
+    click.echo(str(summary))
 
-    click.secho("\nActivity Query\n", fg="green")
-    df = query(ACTIVITIES_QUERY + "\nLIMIT 5", version=version)
-    click.echo(df.to_markdown(index=False))
+    click.secho("ID to Name Query", fg="green")
+    try:
+        df = query(ID_NAME_QUERY + "\nLIMIT 5", version=version)
+    except OSError:
+        click.secho("failed", fg="red")
+    else:
+        click.echo("\n" + df.to_markdown(index=False))
+
+    click.secho("\nActivity Query", fg="green")
+    try:
+        df = query(ACTIVITIES_QUERY + "\nLIMIT 5", version=version)
+    except OSError:
+        click.secho("failed", fg="red")
+    else:
+        click.echo("\n" + df.to_markdown(index=False))
 
 
 @main.command()
@@ -80,41 +86,85 @@ def substructure(version: str | None) -> None:
     get_substructure_library(version=version)
 
 
+TEST_VERSIONS = ["1", "19", "25", "35"]
+
+
 @main.command()
-@click.option("--delete-old", is_flag=True)
-def history(delete_old: bool) -> None:
-    """Generate a history command."""
+@click.option(
+    "--delete-old",
+    is_flag=True,
+    help="Delete old versions as you go, useful if you have limited hard disk "
+    "space but also more time intensive on re-run.",
+)
[…]
