### 2025-08-27 #8844 @dhimmel
I've completed my review. This is a nice brief description of the utility in the paper and the package is useful. All looks good to me.

One tiny suggestion is to consider linking to notebooks in the manuscript using a tag (e.g. `joss`) rather than the `main` branch to help prevent link rot if repository is reorganized. This is not a strong preference and its minor.
