### 2025-08-27 #8844 @cthoyt
That's a good idea @dhimmel, I will do that as a last step in case there are any other software changes - I expect to make a 0.5.2 release concurrent with the end of review, and I update the links in the manuscript accordingly. Thanks!
