+++ README.md
@@ -306,45 +306,45 @@ table can be generated with
 `chembl_downloader history && chembl_downloader history-draw`. These data can be
 found as a TSV in [docs/\_data/summary.tsv](docs/_data/summary.tsv).
 
-| Version | Date       | Compounds |    Assays | Activities | Named Compounds |
-| ------- | ---------- | --------: | --------: | ---------: | --------------: |
-| 35      | 2024-12-01 | 2,496,335 | 1,740,546 | 21,123,501 |          42,231 |
-| 34      | 2024-03-28 | 2,431,025 | 1,644,390 | 20,772,701 |          42,387 |
-| 33      | 2023-05-31 | 2,399,743 | 1,610,596 | 20,334,684 |          41,923 |
-| 32      | 2023-01-26 | 2,354,965 | 1,536,903 | 20,038,828 |          41,923 |
-| 31      | 2022-07-12 | 2,331,700 | 1,498,681 | 19,780,369 |          41,585 |
-| 30      | 2022-02-22 | 2,157,379 | 1,458,215 | 19,286,751 |          41,549 |
-| 29      | 2021-07-01 | 2,105,464 | 1,383,553 | 18,635,916 |          41,383 |
-| 28      | 2021-01-15 | 2,086,898 | 1,358,549 | 17,276,334 |          41,049 |
-| 27      | 2020-05-18 | 1,961,462 | 1,221,361 | 16,066,124 |          40,834 |
-| 26      | 2020-02-14 | 1,950,765 | 1,221,311 | 15,996,368 |          40,822 |
-| 25      | 2019-02-01 | 1,879,206 | 1,125,387 | 15,504,603 |          39,885 |
-| 24.1    | 2018-05-01 | 1,828,820 | 1,060,283 | 15,207,914 |          39,877 |
-| 24      | 2018-05-01 | 1,828,820 | 1,060,283 | 15,207,914 |          39,877 |
-| 23      | 2017-05-18 | 1,735,442 | 1,302,147 | 14,675,320 |          39,584 |
-| 22.1    | 2016-11-17 | 1,686,695 | 1,246,683 | 14,371,197 |          39,422 |
-| 22      | 2016-09-28 | 1,686,695 | 1,246,132 | 14,371,219 |          39,422 |
-| 21      | 2015-02-12 | 1,592,191 | 1,212,831 | 13,968,617 |          39,347 |
-| 20      | 2015-02-03 | 1,463,270 | 1,148,942 | 13,520,737 |          39,016 |
-| 19      | 2014-07-23 | 1,411,786 | 1,106,285 | 12,843,338 |          38,910 |
-| 18      | 2014-04-02 | 1,359,508 | 1,042,374 | 12,419,715 |          35,817 |
-| 17      | 2013-09-16 | 1,324,941 |   734,201 | 12,077,491 |          32,692 |
-| 16      | 2013-05-15 | 1,295,510 |   712,836 | 11,420,351 |          23,532 |
-| 15      | 2013-01-30 | 1,254,575 |   679,259 | 10,509,572 |          23,528 |
-| 14      | 2012-07-18 | 1,213,242 |   644,734 | 10,129,256 |          16,573 |
-| 13      | 2012-02-29 | 1,143,682 |   617,681 |  6,933,068 |          16,397 |
-| 12      | 2011-11-30 | 1,077,189 |   596,122 |  5,654,847 |          16,658 |
-| 11      | 2011-06-07 | 1,060,258 |   582,982 |  5,479,146 |          16,264 |
-| 10      | 2011-06-07 | 1,000,468 |   534,391 |  4,668,202 |          16,159 |
-| 9       | 2011-01-04 |   658,075 |   499,867 |  3,030,317 |           3,746 |
[…]
