+++ README.md
@@ -336,7 +336,7 @@ found as a TSV in [docs/\_data/summary.tsv](docs/_data/summary.tsv).
 |      12 | 2011-11-30 | 1,077,189 |          16,658 |   596,122 |  5,654,847 |    43,418 |   8,703 |     0 |       0 |             0 |                0 |               0 |
 |      11 | 2011-06-07 | 1,060,258 |          16,264 |   582,982 |  5,479,146 |    42,516 |   8,603 |     0 |       0 |             0 |                0 |               0 |
 |      10 | 2011-06-07 | 1,000,468 |          16,159 |   534,391 |  4,668,202 |    40,624 |   8,372 |     0 |       0 |             0 |                0 |               0 |
-|       9 | 2011-01-04 |         0 |           3,746 |   499,867 |  3,030,317 |    39,094 |   8,091 |     0 |       0 |             0 |                0 |               0 |
+|       9 | 2011-01-04 |   658,075 |           3,746 |   499,867 |  3,030,317 |    39,094 |   8,091 |     0 |       0 |             0 |                0 |               0 |
 |       8 | 2010-11-05 |   636,269 |               0 |   488,898 |  2,973,034 |    38,462 |   8,088 |     0 |       0 |             0 |                0 |               0 |
 |       7 | 2010-09-03 |   602,500 |               0 |   485,095 |  2,948,069 |    38,204 |   8,078 |     0 |       0 |             0 |                0 |               0 |
 |       6 | 2010-09-03 |   600,625 |               0 |   481,752 |  2,925,588 |    38,029 |   8,054 |     0 |       0 |             0 |                0 |               0 |
