### 2025-09-04 3ff8791
Pin jupyter notebook links to release version
Suggested in https://github.com/openjournals/joss-reviews/issues/8844#issuecomment-3226500108
