### 2025-09-04 a9f9327
Bump version: 0.5.2-dev → 0.5.2
