### 2025-09-04 b7fcf79
Bump version: 0.5.2 → 0.5.3-dev
