+++ docs/source/conf.py
@@ -24,7 +24,7 @@ copyright = f"{date.today().year}, Charles Tapley Hoyt"
 author = "Charles Tapley Hoyt"
 
 # The full version, including alpha/beta/rc tags.
-release = "0.5.2"
+release = "0.5.3-dev"
 
 # The short X.Y version.
 parsed_version = re.match(
