### 2025-09-04 #8844 @cthoyt
@espottesmith I've done the following:

- Made a v0.5.2 release corresponding to the publication. The version-specific DOI on Zenodo for this release is: https://doi.org/10.5281/zenodo.17053149.
- Simultaneously, pinned the notebook URLs to the tag for the new release as suggested by Daniel above (https://github.com/cthoyt/chembl-downloader/commit/3ff87919216923da903ec9bd04733a068c9e2743)
- Double checked my ORCID and ROR are correct
- Checked the repository and the zenodo record have the same license (MIT)
- Checked the author list is correct on the manuscript, CITATIONS.cff file, and in the Zenodo record (it's just me)

should be good to go
