### 2025-09-04 #8844 @espottesmith
Comments on your manuscript:

- Given that we link the Zenodo archive in the margin of the first page, it isn't necessary to include it in your "Availability and usage" section. You can, of course, leave it there if you want.
- Your "Case studies" section is awesome. This is clearly a valuable community tool!
- Though one could say that the need for this package is evident in its many applications, I nonetheless feel like the "State of the Field" as it presently exists is somewhat lacking. What makes `KNIME` and `Pipeline Pilot` less flexible and extensible than your code? What makes the existing API inefficient? Is it just that the API doesn't go through the FTP server? A small amount of additional detail would be very welcome.

Great work on this! I'm looking forward to being able to accept your manuscript in JOSS.
