### 2025-09-04 6c569c4
Remove redundant archive link in JOSS manuscript
Suggested in https://github.com/openjournals/joss-reviews/issues/8844#issuecomment-3252917890
