### 2025-09-10 982c85e
Update paper.md
