+++ paper/paper.md
@@ -59,9 +59,9 @@ for querying and processing data in bulk due to dependency on network
 connection, remote server uptime and load, rate limits, and the need to paginate
 over results. Alternatively, bulk access is currently cumbersome for most
 potential users due to the need to download, set up, and connect to databases
-locally. Additional third-party software (e.g., KNIME, Pipeline Pilot) provide
-alternate access to ChEMBL, but are relatively inflexible, inextensible, or
-proprietary.
+locally. Finally, third-party software (e.g., Pipeline Pilot, KNIME) that
+provide access to ChEMBL are often inflexible or inextensible due to being
+proprietary, closed source, or lacking approachable documentation.
 
 # Summary
 
