### 2025-09-10 39f1a36
Update paper.md
