+++ paper/paper.md
@@ -59,8 +59,8 @@ for querying and processing data in bulk due to dependency on network
 connection, remote server uptime and load, rate limits, and the need to paginate
 over results. Alternatively, bulk access is currently cumbersome for most
 potential users due to the need to download, set up, and connect to databases
-locally. Finally, third-party software (e.g., Pipeline Pilot, KNIME) that
-provide access to ChEMBL are often inflexible or inextensible due to being
+locally. Finally, third-party software (e.g., Pipeline Pilot, KNIME, Galaxy)
+that provide access to ChEMBL are often inflexible or inextensible due to being
 proprietary, closed source, or lacking approachable documentation.
 
 # Summary
