### 2025-09-10 f989eb6
Improve "state of the field" section (#32)
