### 2025-09-10 9da7395
Add missing references
