+++ paper/paper.bib
@@ -304,49 +304,44 @@
   volume       = {3},
 }
 
-
 @article{Berthold2009,
- author = {Berthold, Michael R. and Cebron, Nicolas and Dill, Fabian and Gabriel, Thomas R. and K\"{o}tter, Tobias and Meinl, Thorsten and Ohl, Peter and Thiel, Kilian and Wiswedel, Bernd},
- title = {KNIME - the Konstanz Information Miner: Version 2.0 and Beyond},
- journal = {SIGKDD Explor. Newsl.},
- issue_date = {June 2009},
- volume = {11},
- number = {1},
- month = nov,
- year = {2009},
- issn = {1931-0145},
- pages = {26--31},
- numpages = {6},
- url = {http://doi.acm.org/10.1145/1656274.1656280},
- doi = {10.1145/1656274.1656280},
- acmid = {1656280},
- publisher = {ACM},
- address = {New York, NY, USA},
+  author       = {Berthold, Michael R. and Cebron, Nicolas and Dill, Fabian and Gabriel, Thomas R. and Kötter, Tobias and Meinl, Thorsten and Ohl, Peter and Thiel, Kilian and Wiswedel, Bernd},
+  location     = {New York, NY, USA},
+  publisher    = {ACM},
+  url          = {http://doi.acm.org/10.1145/1656274.1656280},
+  date         = {2009-11},
+  doi          = {10.1145/1656274.1656280},
+  issn         = {1931-0145},
+  journaltitle = {SIGKDD Explor. Newsl.},
+  number       = {1},
+  pages        = {26--31},
+  title        = {KNIME - the Konstanz Information Miner: Version 2.0 and Beyond},
+  volume       = {11},
 }
 
 @article{Galaxy2024,
-  title = {The {Galaxy} platform for accessible, reproducible, and collaborative data analyses: 2024 update},
-  author = {The Galaxy Community},
-  journal = {Nucleic Acids Res.},
-  doi = {10.1093/nar/gkae410},
-  volume = {52},
-  number = {W1},
-  pages = {W83-W94},
-  year = {2024},
-  url = {https://doi.org/10.1093/nar/gkae410},
-  publisher = {Oxford University Press}
+  author       = {Community, The Galaxy},
+  publisher    = {Oxford University Press},
+  url          = {https://doi.org/10.1093/nar/gkae410},
+  date         = {2024},
+  doi          = {10.1093/nar/gkae410},
+  journaltitle = {Nucleic Acids Res.},
+  number       = {W1},
+  pages        = {W83--W94},
+  title        = {The {Galaxy} platform for accessible, reproducible, and collaborative data analyses: 2024 update},
+  volume       = {52},
 }
 
 @article{Warr2012,
-	title = {Scientific workflow systems: {Pipeline} {Pilot} and {KNIME}},
-	volume = {26},
-	issn = {1573-4951},
-	url = {https://doi.org/10.1007/s10822-012-9577-7},
-	doi = {10.1007/s10822-012-9577-7},
-	number = {7},
-	journal = {Journal of Computer-Aided Molecular Design},
-	author = {Warr, Wendy A.},
-	month = jul,
-	year = {2012},
-	pages = {801--804},
+  author       = {Warr, Wendy A.},
+  url          = {https://doi.org/10.1007/s10822-012-9577-7},
+  date         = {2012-07},
+  doi          = {10.1007/s10822-012-9577-7},
+  issn         = {1573-4951},
+  journaltitle = {Journal of Computer-Aided Molecular Design},
+  number       = {7},
+  pages        = {801--804},
+  title        = {Scientific workflow systems: {Pipeline} {Pilot} and {KNIME}},
+  volume       = {26},
 }
+
