+++ paper/paper.bib
@@ -140,7 +140,6 @@
   publisher    = {Springer International Publishing},
   date         = {2017},
   doi          = {10.1186/s13321-017-0203-5},
-  isbn         = {1758-2946},
   issn         = {17582946},
   journaltitle = {J. Cheminform.},
   number       = {1},
@@ -320,7 +319,7 @@
 }
 
 @article{Galaxy2024,
-  author       = {Community, The Galaxy},
+  author       = {{The Galaxy Community}},
   publisher    = {Oxford University Press},
   url          = {https://doi.org/10.1093/nar/gkae410},
   date         = {2024},
