### 2025-09-10 8de4099
Add missing references for JOSS paper (#33)
