+++ paper/paper.bib
@@ -140,7 +140,6 @@
   publisher    = {Springer International Publishing},
   date         = {2017},
   doi          = {10.1186/s13321-017-0203-5},
-  isbn         = {1758-2946},
   issn         = {17582946},
   journaltitle = {J. Cheminform.},
   number       = {1},
@@ -302,5 +301,46 @@
   pages        = {100069},
   title        = {Pharmaceutical patent landscaping: A novel approach to understand patents from the drug discovery perspective},
   volume       = {3},
+}
+
+@article{Berthold2009,
+  author       = {Berthold, Michael R. and Cebron, Nicolas and Dill, Fabian and Gabriel, Thomas R. and Kötter, Tobias and Meinl, Thorsten and Ohl, Peter and Thiel, Kilian and Wiswedel, Bernd},
+  location     = {New York, NY, USA},
+  publisher    = {ACM},
+  url          = {http://doi.acm.org/10.1145/1656274.1656280},
+  date         = {2009-11},
+  doi          = {10.1145/1656274.1656280},
+  issn         = {1931-0145},
+  journaltitle = {SIGKDD Explor. Newsl.},
+  number       = {1},
+  pages        = {26--31},
+  title        = {KNIME - the Konstanz Information Miner: Version 2.0 and Beyond},
+  volume       = {11},
+}
+
+@article{Galaxy2024,
+  author       = {{The Galaxy Community}},
+  publisher    = {Oxford University Press},
+  url          = {https://doi.org/10.1093/nar/gkae410},
+  date         = {2024},
+  doi          = {10.1093/nar/gkae410},
+  journaltitle = {Nucleic Acids Res.},
+  number       = {W1},
+  pages        = {W83--W94},
+  title        = {The {Galaxy} platform for accessible, reproducible, and collaborative data analyses: 2024 update},
+  volume       = {52},
+}
+
+@article{Warr2012,
+  author       = {Warr, Wendy A.},
+  url          = {https://doi.org/10.1007/s10822-012-9577-7},
+  date         = {2012-07},
+  doi          = {10.1007/s10822-012-9577-7},
+  issn         = {1573-4951},
+  journaltitle = {Journal of Computer-Aided Molecular Design},
+  number       = {7},
+  pages        = {801--804},
+  title        = {Scientific workflow systems: {Pipeline} {Pilot} and {KNIME}},
+  volume       = {26},
 }
 
