### 2025-09-10 #8844 @cthoyt
@espottesmith thanks for the catch, done in https://github.com/cthoyt/chembl-downloader/pull/33
