### 2025-09-10 #8844 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
