### 2025-09-12 4210c17
Add city and country to affiliation
