### 2025-09-13 #8844 @cthoyt
@Kevin-Mattheus-Moerman @espottesmith @dhimmel @PatWalters a big thanks to all of you for the review and editorial work. I appreciate it very much! I just checked through everything and the issue is ready to close.
